quickbite/backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=21.2.0
whitenoise>=6.6.0
